+++ screens/mycomponents/entryDynamicButton.tsx
