+++ screens/mycomponents/resultsView.tsx
