+++ screens/mycomponents/timer.tsx
@@ -30,8 +30,8 @@ const timerPresentation = props => {
             time == 0
               ? styles.timerRed
               : time < 10
-              ? styles.timerYellow
-              : styles.timerGreen
+                ? styles.timerYellow
+                : styles.timerGreen
           }
         />
       </View>
