+++ screens/mycomponents/timerOptions.tsx
